feat: Introduce new query and result types for batch and failed event handling, and update endpoints to utilize them
