--- ActionProcessor/Application/Commands/FileCommands.cs
@@ -1,27 +1,0 @@
-namespace ActionProcessor.Application.Commands;
-// TODO: Separar os commands
-public record UploadFileCommand(
-    IFormFile File,
-    string UserEmail,
-    string? SideEffects = null
-);
-
-public record UploadFileResult(
-    Guid BatchId,
-    string FileName,
-    int TotalEvents,
-    bool Success,
-    string? ErrorMessage = null
-);
-
-public record RetryFailedEventsCommand(
-    Guid BatchId,
-    IEnumerable<Guid>? EventIds = null,
-    string UserEmail = ""
-);
-
-public record RetryFailedEventsResult(
-    int EventsRetried,
-    bool Success,
-    string? ErrorMessage = null
-);
