+++ ActionProcessor/Application/Commands/RetryFailedEventsCommand.cs
@@ -1,0 +1,7 @@
+namespace ActionProcessor.Application.Commands;
+
+public sealed record RetryFailedEventsCommand(
+    Guid BatchId,
+    IEnumerable<Guid>? EventIds = null,
+    string UserEmail = ""
+);
