+++ ActionProcessor/Application/Commands/UploadFileCommand.cs
@@ -1,0 +1,7 @@
+namespace ActionProcessor.Application.Commands;
+
+public sealed record UploadFileCommand(
+    IFormFile File,
+    string UserEmail,
+    string? SideEffects = null
+);
