+++ ActionProcessor/Application/Handlers/GetBatchStatusQueryHandler.cs
@@ -1,4 +1,5 @@
 using ActionProcessor.Application.Queries;
+using ActionProcessor.Application.Results;
 using ActionProcessor.Domain.Interfaces;
 
 namespace ActionProcessor.Application.Handlers;
