+++ ActionProcessor/Application/Handlers/GetFailedEventsQueryHandler.cs
@@ -1,4 +1,5 @@
 using ActionProcessor.Application.Queries;
+using ActionProcessor.Application.Results;
 using ActionProcessor.Domain.Entities;
 using ActionProcessor.Domain.Interfaces;
