+++ ActionProcessor/Application/Handlers/RetryEventsFailedCommandHandler.cs
@@ -1,4 +1,5 @@
 using ActionProcessor.Application.Commands;
+using ActionProcessor.Application.Results;
 using ActionProcessor.Domain.Entities;
 using ActionProcessor.Domain.Interfaces;
