+++ ActionProcessor/Application/Queries/GetBatchListQuery.cs
@@ -1,0 +1,7 @@
+namespace ActionProcessor.Application.Queries;
+
+public sealed record GetBatchListQuery(
+    int Skip = 0,
+    int Take = 100,
+    string? UserEmail = null
+);
