+++ ActionProcessor/Application/Queries/GetBatchStatusQuery.cs
@@ -1,0 +1,3 @@
+namespace ActionProcessor.Application.Queries;
+
+public sealed record GetBatchStatusQuery(Guid BatchId);
