+++ ActionProcessor/Application/Queries/GetFailedEventsQuery.cs
@@ -1,0 +1,6 @@
+namespace ActionProcessor.Application.Queries;
+
+public sealed record GetFailedEventsQuery(
+    Guid? BatchId = null,
+    string? UserEmail = null
+);
