+++ ActionProcessor/Application/Results/RetryFailedEventsResult.cs
@@ -1,0 +1,7 @@
+namespace ActionProcessor.Application.Results;
+
+public sealed record RetryFailedEventsResult(
+    int EventsRetried,
+    bool Success,
+    string? ErrorMessage = null
+);
