+++ ActionProcessor/Application/Results/UploadFileResult.cs
@@ -1,0 +1,9 @@
+namespace ActionProcessor.Application.Results;
+
+public sealed record UploadFileResult(
+    Guid BatchId,
+    string FileName,
+    int TotalEvents,
+    bool Success,
+    string? ErrorMessage = null
+);
